Improve formatting and refactor code
