+++ docs/SUMMARY.md
@@ -3,6 +3,7 @@
 - [Overview](./overview.md)
 - [Using the source generator](./generator/generator.md)
     - [Configuration options](./generator/options.md)
+- [Data model](./data-model.md)
 - [Customization](./customization.md)
 - [External types](./foreign-types.md)
 - [Generic types](./generic-types.md)
