+++ docs/book.toml
@@ -3,4 +3,4 @@ authors = ["Andy Gocke"]
 language = "en"
 multilingual = false
 src = "."
-title = "serde-dn"
+title = "Serde.NET"
