+++ src/serde/ISerde.cs
@@ -11,6 +11,16 @@ public interface ISerde<T> : ISerialize<T>, IDeserialize<T> { }
 /// Convenience interface for implementing both serialization and deserialization providers.
 /// See <see cref="ISerializeProvider{T}"/> and <see cref="IDeserializeProvider{T}"/> for details.
 /// </summary>
+public interface ISerdeProvider<T> : ISerializeProvider<T>, IDeserializeProvider<T>
+{ }
+
+/// <summary>
+/// Convenience interface for implementing both serialization and deserialization providers.
+/// See <see cref="ISerializeProvider{T}"/> and <see cref="IDeserializeProvider{T}"/> for details.
+///
+/// This version supports producing a singleton <see cref="Instance"/> of the implementing type, and
+/// uses that instance for both serialization and deserialization.
+/// </summary>
 public interface ISerdeProvider<TSelf, TSerde, T> : ISerializeProvider<T>, IDeserializeProvider<T>
     where TSelf : ISerdeProvider<TSelf, TSerde, T>
     where TSerde : ISerde<T>
@@ -22,10 +32,4 @@ public interface ISerdeProvider<TSelf, TSerde, T> : ISerializeProvider<T>, IDese
 
 public interface ISerdeProvider<TSelf, T> : ISerdeProvider<TSelf, ISerde<T>, T>
     where TSelf : ISerdeProvider<TSelf, T>, ISerde<T>
-{ }
-
-/// <summary>
-/// Provider interface for retrieving the serialization and deserialization objects for a type.
-/// </summary>
-public interface ISerdeProvider<T> : ISerializeProvider<T>, IDeserializeProvider<T>
 { }
